move Footer css to Footer.module.css
