+++ src/App.jsx
@@ -6,7 +6,7 @@ import "./App.css"
 import { AiOutlineReload } from "react-icons/ai"
 import ColorInput from "./components/ColorInput/ColorInput"
 import ContrastInfoCard from "./components/ContrastInfoCard/ContrastInfoCard"
-import Footer from "./components/Footer"
+import Footer from "./components/Footer/Footer"
 
 const initialColorState = [
 	{ colorCodeHex: "#ffffff", relativeLuminance: 1 },
