+++ src/components/Footer/Footer.module.css
@@ -1,0 +1,21 @@
+
+/* Footer Component */
+.footer {
+	margin-top: auto;
+	padding: 5rem 0 1rem 0;
+	display: flex;
+	justify-content: space-between;
+	font-family: inherit;
+}
+
+.footerDiv {
+	display: flex;
+	justify-content: center;
+	align-items: center;
+	gap: 0.5rem;
+}
+
+.footerDiv svg {
+	width: 1.2rem;
+	height: auto;
+}
